requirements.txt
streamlit
opencv-python
numpy
torch
torchvision
Pillow
matplotlib
ultralytics